requirements.txt
APScheduler==3.5.3
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysql-connector-python==8.0.12
protobuf==3.6.1
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
tzlocal==1.5.1
Werkzeug==0.14.1
